packages/backend/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# API utilities
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/LLM integrations
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.4.0

# Monitoring and logging
prometheus-client==0.20.0
structlog==24.1.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0

# Development
black==24.2.0
ruff==0.3.0
mypy==1.8.0